# Do not Clone `PathElem`s when creating a map value's path.